K8s-project/backend/requirements.txt
flask
flask-cors
redis
psycopg2-binary
